rename connection-pool -> dbConnection
