+++ hrm-domain/hrm-core/case/caseDataAccess.ts
@@ -14,7 +14,7 @@
  * along with this program.  If not, see https://www.gnu.org/licenses/.
  */
 
-import { db, pgp } from '../connection-pool';
+import { db, pgp } from '../dbConnection';
 import { getPaginationElements } from '../search';
 import { PATCH_CASE_INFO_BY_ID, updateByIdSql } from './sql/caseUpdateSql';
 import {
