+++ hrm-domain/hrm-core/conversation-media/sql/conversation-media-insert-sql.ts
@@ -14,7 +14,7 @@
  * along with this program.  If not, see https://www.gnu.org/licenses/.
  */
 
-import { pgp } from '../../connection-pool';
+import { pgp } from '../../dbConnection';
 import type { NewConversationMedia } from '../conversation-media-data-access';
 import { HrmAccountId } from '@tech-matters/types';
