+++ hrm-domain/hrm-core/profile/sql/profile-list-sql.ts
@@ -14,7 +14,7 @@
  * along with this program.  If not, see https://www.gnu.org/licenses/.
  */
 
-import { pgp } from '../../connection-pool';
+import { pgp } from '../../dbConnection';
 import {
   DateExistsCondition,
   DateFilter,
